+++ lib/core/asset_workers/public_key_worker/stub_ssh_handler.ex
@@ -1,5 +1,0 @@
-defmodule FarmbotOS.PublicKeyHandler.StubSSHHandler do
-  @behaviour FarmbotOS.Asset.PublicKey
-  def ready?(), do: true
-  def add_key(_key), do: :ok
-end