+++ CHANGELOG.md
@@ -5,6 +5,8 @@
  * Ability to manage `resource` variables from Lua.
  * Upgrade Nerves system, Erlang, Elixir.
  * Ability to specify `safe_z` from Lua `move_absolute()`.
+ * Remove SSH subsystem. It is no longer maintained.
+ * Nerves system upgrades
 
 # 14.8.1