+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 14.9.0
 
  * Ability to manage `resource` variables from Lua.
+ * Ability to execute GCode raw from Lua.
  * Upgrade Nerves system, Erlang, Elixir.
  * Ability to specify `safe_z` from Lua `move_absolute()`.
  * Remove SSH subsystem. It is no longer maintained.